Calculation of the predictive variable influence on projection

Calculation of the predictive variable influence on projection

Usage

calPredVIP(SSXAP, SSYAP, SSD, loading)

Arguments

SSXAP

the sum of squares values of deflated X matrix for the predictive VIPO2PLS

SSYAP

the sum of squares values of deflated Y matrix for the predictive VIPO2PLS

SSD

the sum of square values

loading

the normalized loading matrices